refactor: フロントエンド Badge 集約 + 型定義整理 + Dashboard 分割
ScoreBadge/StatusBadge/SeverityBadge を ui/Badge.tsx に集約し、
AnalysisData を types/index.ts に追加。Dashboard.tsx を
EpisodeCreationForm + EpisodeTable に分割（245行 → 24行）。
