+++ backend/app/api/schemas.py
@@ -4,7 +4,6 @@ from datetime import datetime
 
 from pydantic import BaseModel
 
-
 # --- Request schemas ---
